seperate to components
