+++ src/components/recipe.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Recipe = ({ recipe }) => (
+  <li>{ recipe }</li>
+);
+
+export default Recipe;
